+++ docsource/releasenotes.rst
@@ -24,6 +24,7 @@ NEW FEATURES:
   ``deploy.cfg.example`` for an example). See
   ``us.kbase.workspace.test.listener.NullListenerFactory`` for an example implementation.
 * Added the ``list_workspace_ids`` method.
+* Added the ``listWorkspaceIDs`` administration command.
 * Added the ``getPermissionsMass`` administration command.
 * Added the ``getWorkspaceInfo`` administration command.
 * Added the ``listObjects`` administration command.